requirements.txt
certifi==2018.8.24
chardet==3.0.4
Django==2.1.1
djangorestframework==3.8.2
httpie==0.9.9
idna==2.7
psycopg2==2.7.5
psycopg2-binary==2.7.5
Pygments==2.2.0
pytz==2018.5
requests==2.19.1
urllib3==1.23
